@@ -1,6 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    images: {
+
+  images: {
         remotePatterns: [
           {
             protocol: 'https',
@@ -12,7 +13,8 @@ const nextConfig = {
           }
         ],
       },
-   
-}
 
-module.exports = nextConfig
+
+};
+
+export default nextConfig;
