@@ -1,1 +1,3 @@
-export { GET, POST } from "@/auth"
+export { /* @next-codemod-error `GET` export is re-exported. Check if this component uses `params` or `searchParams`*/
+GET, /* @next-codemod-error `POST` export is re-exported. Check if this component uses `params` or `searchParams`*/
+POST } from "@/auth"
